Add files via upload
I changed the menu so that it now has two options. One option will allow a user to sign in and the other allows them to create a new account. When a user chooses to create an account, it will first ask for a user name. The system ensures that no other user has that name, ask and confirms a new password, then allows the user to input information regarding their classes. The program takes the user through every semester allowing them to add information like the semester name (such as Spring 2020), class name, class grade, class weight, and ect. When the user finishes typing in their information, it take them to a second menu. If a user instead selects the option to sign in, it will ask them for a username and a password. After the user inputs the information, and the program validates it, the program will load in a text file with the user's information and then take them to another menu. The second menu has a total of five options:
Print off a semester -
List the semester name, how many classes are in the semester, class names, grades, and the semester GPA.
Print entire academic career -
List all the semesters followed by the overall GPA.
Add a new Semester
Print off Clemson GPA
Print off Life GPA}
Once the user is done with the program, they can select to exit and all of their information is stored into a text file for the next time they sign in.
**Make sure you add the textfiles into a new folder named "UserFiles" or it won't work.